+++ src/lib/getPeerDependenciesFromRegistry.ts
@@ -2,9 +2,45 @@ import ProgressBar from 'progress'
 import nodeSemver from 'semver'
 import { Index } from '../types/IndexType'
 import { Options } from '../types/Options'
-import { VersionSpec } from '../types/VersionSpec'
+import { Version } from '../types/Version'
 import getPackageManager from './getPackageManager'
-import { isCircularPeer } from './isCircularPeer'
+
+type CircularData =
+  | {
+      isCircular: true
+      offendingPackage: string
+    }
+  | {
+      isCircular: false
+    }
+
+/**
+ * Checks if the specified package will create a loop of peer dependencies by traversing all paths to find a cycle
+ *
+ * If a cycle was found, the offending peer dependency of the specified package is returned
+ */
+function isCircularPeer(peerDependencies: Index<Index<string>>, packageName: string): CircularData {
+  let queue = [[packageName]]
+  while (queue.length > 0) {
+    const nextQueue: string[][] = []
+    for (const path of queue) {
+      const parents = Object.keys(peerDependencies[path[0]] ?? {})
+      for (const name of parents) {
+        if (name === path.at(-1)) {
+          return {
+            isCircular: true,
+            offendingPackage: path[0],
+          }
+        }
+        nextQueue.push([name, ...path])
+      }
+    }
+    queue = nextQueue
+  }
+  return {
+    isCircular: false,
+  }
+}
 
 /**
  * Get the latest or greatest versions from the NPM repository based on the version target.
@@ -13,7 +49,7 @@ import { isCircularPeer } from './isCircularPeer'
  * @param [options={}] Options.
  * @returns Promised {packageName: peer dependencies} collection
  */
-async function getPeerDependenciesFromRegistry(packageMap: Index<VersionSpec>, options: Options) {
+async function getPeerDependenciesFromRegistry(packageMap: Index<Version>, options: Options) {
   const packageManager = getPackageManager(options, options.packageManager)
   if (!packageManager.getPeerDependencies) return {}
